Update readme to indicate this is 'the hard way'
This page has high google-juice for 'kubernetes cluster federation'.  It's incredibly useful to have as a learning resource but now we have `kubefed` on the way, it's not necessarily the first way that someone should solve this problem.

The Kubernetes The Hard Way tutorial makes this distinction very clearly, so I've lifted some of its readme and modified it to fit here.